storage changes

+++ src/kernel/include/drivers/storage/ahci.hpp
@@ -233,8 +233,8 @@ class ahci_storage_driver_t : public storage_driver_interface_t {
 public:
     ahci_storage_driver_t(ahci_drive_t* drive);
 
-    bool read(uint32_t lba, uint8_t* buffer, size_t size) override;
-    bool write(uint32_t lba, uint8_t* buffer, size_t size) override;
+    bool read(uint64_t lba, uint8_t* buffer, size_t size) override;
+    bool write(uint64_t lba, uint8_t* buffer, size_t size) override;
     size_t get_block_size() override;
     void set_root_lba(uint64_t lba) override;
     uint64_t get_root_lba() override;
